+++ module-system/module-implementations/sov-blob-storage/tests/integration/blob_storage_tests.rs
@@ -1,7 +1,7 @@
 use sov_blob_storage::BlobStorage;
 use sov_chain_state::{ChainState, ChainStateConfig};
 use sov_mock_da::{MockAddress, MockDaSpec};
-use sov_modules_api::batch::BatchWithId;
+use sov_modules_api::batch::{Batch, BatchWithId};
 use sov_modules_api::runtime::capabilities::RawTx;
 use sov_modules_api::{KernelModule, KernelWorkingSet, StateCheckpoint};
 use sov_prover_storage_manager::new_orphan_storage;
@@ -74,10 +74,12 @@ fn store_and_retrieve_standard() {
 
     let mut batches = Vec::new();
     for i in 1..=5 {
+        let txs = vec![RawTx {
+            data: vec![i * 3 + 1, i * 3 + 2, i * 3 + 3],
+        }];
+
         let batch = BatchWithId {
-            txs: vec![RawTx {
-                data: vec![i * 3 + 1, i * 3 + 2, i * 3 + 3],
-            }],
+            batch: Batch { txs },
             id: [i; 32],
         };
         batches.push((batch, sender));